use isPrivate in MinterAccessControl
