+++ tokens/test/erc-721/Rarible.test.js
@@ -396,83 +396,6 @@ contract("ERC721Rarible", accounts => {
     );
   });
 
-  it("mint and transfer with minter access control", async () => {
-    const minter = accounts[1];
-    let transferTo = accounts[2];
-
-    const tokenId = minter + "b00000000000000000000001";
-    const tokenURI = "//uri";
-
-    await token.enableMinterAccessControl({from: tokenOwner});
-    assert.equal(await token.minterAccessControlEnabled(), true);
-
-    await expectThrow(
-      token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [zeroWord]], transferTo, {from: minter})
-    );
-
-    await token.grantMinter(minter, {from: tokenOwner})
-    assert.equal(await token.isValidMinter(minter), true);
-    assert.equal(await token.isValidMinter(transferTo), false);
-
-    await token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [zeroWord]], transferTo, {from: minter})
-    assert.equal(await token.ownerOf(tokenId), transferTo);
-  });
-
-  it("mint and transfer with minter access control and minter signature", async () => {
-    const minter = accounts[1];
-    let transferTo = accounts[2];
-
-    const tokenId = minter + "b00000000000000000000001";
-    const tokenURI = "//uri";
-
-    await token.enableMinterAccessControl({from: tokenOwner});
-    assert.equal(await token.minterAccessControlEnabled(), true);
-
-    const signature = await getSignature(tokenId, tokenURI, creators([minter]), [], minter);
-
-    let whiteListProxy = accounts[5];
-    await token.setDefaultApproval(whiteListProxy, true, {from: tokenOwner});
-
-    await expectThrow(
-      token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [signature]], transferTo, {from: whiteListProxy})
-    );
-
-    await token.grantMinter(minter, {from: tokenOwner})
-    assert.equal(await token.isValidMinter(minter), true);
-    assert.equal(await token.isValidMinter(whiteListProxy), false);
-
-    await token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [signature]], transferTo, {from: whiteListProxy})
-    assert.equal(await token.ownerOf(tokenId), transferTo);
-  });
-
-  it("mint and transfer with minter access control and wrong minter signature", async () => {
-    const minter = accounts[1];
-    let transferTo = accounts[2];
-
-    const tokenId = minter + "b00000000000000000000001";
-    const tokenURI = "//uri";
-
-    await token.enableMinterAccessControl({from: tokenOwner});
-    assert.equal(await token.minterAccessControlEnabled(), true);
-
-    const signature = await getSignature(tokenId, tokenURI, creators([minter]), [], transferTo);
-
-    let whiteListProxy = accounts[5];
-    await token.setDefaultApproval(whiteListProxy, true, {from: tokenOwner});
-
-    await expectThrow(
-      token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [signature]], transferTo, {from: whiteListProxy})
-    );
-
-    await token.grantMinter(minter, {from: tokenOwner})
-    assert.equal(await token.isValidMinter(minter), true);
-    assert.equal(await token.isValidMinter(whiteListProxy), false);
-
-    await expectThrow(
-      token.mintAndTransfer([tokenId, tokenURI, creators([minter]), [], [signature]], transferTo, {from: whiteListProxy})
-    );
-  });
-
   it("standard transfer from owner", async () => {
     let minter = accounts[1];
     const tokenId = minter + "b00000000000000000000001";
